refactoring PutCSV

+++ src/main/java/by/spk/price/entity/Price.java
@@ -5,10 +5,10 @@ import lombok.Data;
 @Data
 public class Price {
 
-    private int brandId;
-    private int categoryId;
-    private int subCategoryId;
-    private int productId;
-    private int recommendedPrice;
+    private Long brandId;
+    private Long categoryId;
+    private Long subCategoryId;
+    private Long productId;
+    private Long recommendedPrice;
 
 }
